Restructuring of the tests and .env location
