+++ tests/conftest.py
@@ -1,4 +1,6 @@
+'''Fixtures for Pytest'''
 from mixer.backend.django import mixer
+from motivator.models import Goal
 from users.models import User, UserMotivator
 import pytest
 
@@ -9,3 +11,10 @@ def user(db):
     test_user = mixer.blend(User, username='Test User')
     motivator.user = test_user
     return motivator
+
+
+@pytest.fixture()
+def goal(user):
+    motivator_goal = mixer.blend(Goal)
+    motivator_goal.user = user.user
+    return motivator_goal
