+++ tests/motivator/test_motivator_views.py
